Move FetchStatsForComplications out of order totals package
