+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/complications/FetchStatsForComplications.kt
@@ -1,9 +1,9 @@
-package com.woocommerce.android.wear.complications.ordertotals
+package com.woocommerce.android.wear.complications
 
 import android.icu.text.CompactDecimalFormat
-import com.woocommerce.android.wear.complications.ordertotals.FetchStatsForComplications.StatType.ORDER_COUNT
-import com.woocommerce.android.wear.complications.ordertotals.FetchStatsForComplications.StatType.ORDER_TOTALS
-import com.woocommerce.android.wear.complications.ordertotals.FetchStatsForComplications.StatType.VISITORS
+import com.woocommerce.android.wear.complications.FetchStatsForComplications.StatType.ORDER_COUNT
+import com.woocommerce.android.wear.complications.FetchStatsForComplications.StatType.ORDER_TOTALS
+import com.woocommerce.android.wear.complications.FetchStatsForComplications.StatType.VISITORS
 import com.woocommerce.android.wear.ui.login.LoginRepository
 import com.woocommerce.android.wear.ui.stats.datasource.StatsRepository
 import kotlinx.coroutines.CoroutineScope
