+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/complications/ordertotals/OrderTotalsComplicationService.kt
@@ -5,8 +5,9 @@ import androidx.wear.watchface.complications.data.ComplicationType
 import androidx.wear.watchface.complications.datasource.ComplicationRequest
 import androidx.wear.watchface.complications.datasource.SuspendingComplicationDataSourceService
 import com.woocommerce.android.R
+import com.woocommerce.android.wear.complications.FetchStatsForComplications
+import com.woocommerce.android.wear.complications.FetchStatsForComplications.StatType.ORDER_TOTALS
 import com.woocommerce.android.wear.complications.createTextComplicationData
-import com.woocommerce.android.wear.complications.ordertotals.FetchStatsForComplications.StatType.ORDER_TOTALS
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
